Address default php.ini setting for php 7.4+
As of PHP 7.4, zend.exception_ignore_args was created and set to 'true'.
This has meant that a feature of SBA is not supported, specifically relating
to the stock of attributed product. This modification checks to see if
that setting is present, if it is present then it attempts to modify the
setting to be 'false', then to return it to its former value.

Fixes #95 at least allows continued operation of the same method of determination.